Move catalog client mock into conftest
