+++ tests/conftest.py
@@ -1,3 +1,4 @@
+from collections.abc import Callable
 from datetime import datetime
 from types import SimpleNamespace
 from uuid import UUID, uuid4
@@ -37,6 +38,59 @@ class FakeTrackerService:
         return SimpleNamespace(status="success")
 
 
+class MockCatalogClient:
+    """Records catalog and tracker health-check calls for service-list tests."""
+
+    def __init__(self) -> None:
+        self.headers: object = None
+        self.get_url: str | None = None
+        self.post_url: str | None = None
+        self.json: dict[str, object] | None = None
+
+    def get(self, url: str, *, params: dict[str, object] | None = None) -> httpx.Response:
+        self.get_url = url
+
+        return httpx.Response(
+            200,
+            json={"services": [{"name": "swebench", "url": "https://swebench.benchmarks.vals.ai/"}]},
+        )
+
+    def post(
+        self,
+        url: str,
+        *,
+        params: dict[str, object] | None = None,
+        json: dict[str, object],
+    ) -> httpx.Response:
+        self.post_url = url
+        self.json = json
+
+        return httpx.Response(
+            200,
+            json={
+                "services": [
+                    {
+                        "name": "swebench",
+                        "url": "https://swebench.benchmarks.vals.ai",
+                        "healthy": True,
+                        "latency_ms": 12,
+                        "error": None,
+                    }
+                ]
+            },
+        )
+
+    def close(self) -> None:
+        pass
+
+
+@pytest.fixture
+def mock_catalog_client_factory() -> Callable[[], MockCatalogClient]:
+    """Build fresh catalog client mocks for service-list tests."""
+
+    return MockCatalogClient
+
+
 @pytest.fixture
 def connect_stream_testbed(monkeypatch: pytest.MonkeyPatch) -> tuple[UUID, list[str]]:
     started_run_id = uuid4()
@@ -58,9 +112,14 @@ def connect_stream_testbed(monkeypatch: pytest.MonkeyPatch) -> tuple[UUID, list[
     def stream_benchmark_status(_tracker: FakeTrackerService, run_id: object) -> None:
         streamed_run_ids.append(str(run_id))
 
+    def check_tracker_service_health(_tracker: FakeTrackerService) -> bool:
+        is_healthy = True
+
+        return is_healthy
+
     monkeypatch.setattr(cli_main, "TrackerService", FakeTrackerService)
     monkeypatch.setattr(cli_main, "get_contract_from_s3", get_contract_from_s3)
-    monkeypatch.setattr(cli_main, "check_tracker_service_health", lambda _tracker: True)
+    monkeypatch.setattr(cli_main, "check_tracker_service_health", check_tracker_service_health)
     monkeypatch.setattr(cli_main, "stream_benchmark_status", stream_benchmark_status)
 
     return started_run_id, streamed_run_ids
